zip/requirements.txt
flask==2.2.3
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
scipy==1.10.1
scikit-image==0.20.0
